% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicParser.R
\name{parseHiC}
\alias{parseHiC}
\title{Parser for data in hic format}
\usage{
parseHiC(paths, replicates, conditions, binSize)
}
\arguments{
\item{paths}{A vector of paths to \code{.hic} files.}

\item{binSize}{The resolution (span of each position in number of bases) to select within
the \code{.hic} files.}

\item{replicates}{A vector of replicate names repeated along the conditions.}

\item{conditions}{A vector of condition names repeated along the replicates.}
}
\value{
An InteractionSet.
}
\description{
Parses interactions in \code{.hic} format and returns an InteractionSet
object.
}
\examples{
\dontrun{
#' # Path to each file
paths <- c(
    "path/to/condition-1.replicate-1.hic",
    "path/to/condition-1.replicate-2.hic",
    "path/to/condition-2.replicate-1.hic",
    "path/to/condition-2.replicate-2.hic",
    "path/to/condition-3.replicate-1.hic"
)

# Replicate and condition of each file. Can be names instead of numbers.
replicates <- c(1, 2, 1, 2, 1)
conditions <- c(1, 1, 2, 2, 3)

# Resolution to select
binSize <- 500000

# Instantiation of data set
hic.experiment <- parseHiC(
    paths,
    replicates = replicates,
    conditions = conditions,
    binSize = binSize
)
}
}
