% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCParser-package.R
\docType{package}
\name{HiCParser-package}
\alias{HiCParser}
\alias{HiCParser-package}
\title{Parse HiC data to import them into R}
\description{
Parser for HiC data in R. TODO To develop. Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emaigne/HiCParser}
  \item Report bugs at \url{https://github.com/emaigne/HiCParser/issues}
}

}
