% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{mergeInteractionSet}
\alias{mergeInteractionSet}
\title{Merge two InteractionSet objects}
\usage{
mergeInteractionSet(interactionSet1, interactionSet2, fill = NA)
}
\arguments{
\item{interactionSet1}{The first \code{\link{InteractionSet}}.}

\item{interactionSet2}{The second \code{\link{InteractionSet}}.}

\item{fill}{Fill missing values with this.}
}
\value{
The merged \code{\link{InteractionSet}}.
}
\description{
Merge two different \code{\link{InteractionSet}}.
}
