% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularParsers.R
\name{parseTabular}
\alias{parseTabular}
\title{Parser for tabular data}
\usage{
parseTabular(path, sep = '\t')
}
\arguments{
\item{sep}{The separator of the tabular file. Default to tabulation.}

\item{path}{A path to a tabular file.}
}
\value{
An InteractionSet object.
}
\description{
Read the file, and fills it using \code{\link{.setFromTabular}}.
}
\examples{
path <- system.file("extdata", "liver_18_10M_500000.tsv", package = "HiCDOC")
object <- parseTabular(path, sep = "\t")

}
