% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularParsers.R
\name{parseTabular}
\alias{parseTabular}
\title{Parser for tabular data}
\usage{
parseTabular(path, sep = '\t')
}
\arguments{
\item{path}{A path to a tabular file.}

\item{sep}{The separator of the tabular file. Default to tabulation.}
}
\value{
An InteractionSet object.
}
\description{
Read the file, produce an \code{InteractionSet} object.
}
\examples{
path <- system.file("extdata", "liver_18_10M_500000.tsv", package = "HiCParser")
object <- parseTabular(path, sep = "\t")

}
